fix:assert folder move splash 폴더 구조 변경했습니다.
