+++ src/components/Template/splash/disk/DiskTemplate.style.tsx
