+++ src/components/Template/splash/disk/index.ts
@@ -1,0 +1,3 @@
+import DiskTemplate from './DiskTemplate';
+
+export default DiskTemplate;
