--- src/components/Template/splash/index.ts
@@ -1,5 +1,0 @@
-import SplashMainTemplate from './SplashMainTemplate';
-import SplashDiskTemplate from './SplashDiskTemplate';
-import SplashMatchedTemplate from './SplashMatchedTemplate';
-
-export { SplashMainTemplate, SplashDiskTemplate, SplashMatchedTemplate };
