+++ src/components/Template/splash/main/MainTemplate.style.tsx
