+++ src/components/Template/splash/main/index.ts
@@ -1,0 +1,3 @@
+import MainTemplate from './MainTemplate';
+
+export default MainTemplate;
