+++ src/components/Template/splash/matched/index.ts
@@ -1,0 +1,3 @@
+import MatchedTemplate from './MatchedTemplate';
+
+export default MatchedTemplate;
